requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
plotly>=5.17.0
numpy>=1.25.2,<2.0.0
scipy>=1.11.0,<2.0.0
transformers==4.35.0
torch==2.6.0
torchvision==0.21.0
Pillow>=10.0.0
requests>=2.31.0
wordcloud>=1.9.2
matplotlib>=3.7.0 